+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ![version 2.1.0](https://img.shields.io/badge/version-2.1.0-brightgreen.svg?style=flat) ![date 2017-03-15](https://img.shields.io/badge/date-2017--03--15-red.svg?style=flat)
 
+**Added:**
+
+- MFF 2.9.5 compliant
+
 **Enhancements:**
 
 - Sort characters list by any attributes