+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased ![version 2.2.0](https://img.shields.io/badge/version-2.2.0-brightgreen.svg?style=flat) ![date 2017-03-15](https://img.shields.io/badge/date-2017--03--15-red.svg?style=flat)
+# Unreleased ![version 2.2.0](https://img.shields.io/badge/version-2.2.0-brightgreen.svg?style=flat) ![date 2017-03-19](https://img.shields.io/badge/date-2017--03--19-red.svg?style=flat)
 
 **In progress**
 
@@ -7,6 +7,11 @@
 **Added:**
 
 - [#16] Keep track of last update and sort by last update
+
+**Fixed bugs:**
+
+- Fix Chrome issue with the search input (Chrome macOs Sierra, I have no clue for windows)
+- [#29] Fix chart when filtering return no matching character
 
 ![version 2.1.0](https://img.shields.io/badge/version-2.1.0-brightgreen.svg?style=flat) ![date 2017-03-15](https://img.shields.io/badge/date-2017--03--15-red.svg?style=flat)